audio_processor/requirements.txt
sounddevice
scipy
numpy
ollama
openai-whisper
requests # For MS Teams webhook
python-dotenv # For managing webhook URL
# For whisper, you might also need ffmpeg installed on the system
# Microsoft Graph API library (will add later if webhook is not sufficient)
